Removing rotation synthesis feature from benchq
