+++ pyproject.toml
@@ -25,8 +25,6 @@ exclude = [
 [[tool.mypy.overrides]]
 module = [
     'juliacall.*',
-    'pyliqtr.*',
-    'pyLIQTR.*',
     'scipy.*',
     'sympy.*',
     'rapidjson',
