+++ setup.cfg
@@ -34,7 +34,6 @@ install_requires =
     numpy>=1.20
     more-itertools~=9.1.0
     pandas==1.5.3
-    pyLIQTR==0.3.3
     openfermion[resources]~=1.6.0
     pytest # required by OpenFermion's resource_estimates module
     graph-state-generation==0.2.0
