+++ src/benchq/algorithms/data_structures/algorithm_implementation.py
@@ -18,15 +18,6 @@ class AlgorithmImplementation:
         program = QuantumProgram.from_circuit(import_circuit(circuit))
         return AlgorithmImplementation(program, error_budget, n_shots)
 
-    def transpile_to_clifford_t(self):
-        return AlgorithmImplementation(
-            self.program.transpile_to_clifford_t(
-                self.error_budget.transpilation_failure_tolerance
-            ),
-            self.error_budget,
-            self.n_shots,
-        )
-
     @property
     def n_t_gates_after_transpilation(self):
         return self.program.get_n_t_gates_after_synthesis(
