+++ src/benchq/compilation/circuits/__init__.py
@@ -1,5 +1,1 @@
 from .compile_to_native_gates import compile_to_native_gates
-from .pyliqtr_transpilation import (
-    get_num_t_gates_per_rotation,
-    pyliqtr_transpile_to_clifford_t,
-)
